restructuring (flat is better than nested, flat is better than nested, flat is be...) ;-)
