+++ obspy/io/nied/fnetmt.py
