+++ obspy/io/nied/knet.py
