+++ obspy/io/nied/util.py
